refactor(FileTransfer): move to lib
